Refactored image grid
